ClassifiedListing removed from ClassifiedListingEvent ClassifiedListingEvent ctor update
